+++ frontend/src/expense/pages/ExpenseMonthlyPage.tsx
@@ -4,8 +4,8 @@ import { useSearchParams } from "react-router"
 import { api } from "../../common/libs/api"
 import { getErrorMessage } from "../../common/libs/client"
 import type { ExpenseResponse } from "../../common/libs/types"
-import { defaultFilter, type ExpenseFilter } from "../components/ExpenseFilterSheet"
 import { ExpenseList } from "../components/ExpenseList"
+import { defaultFilter, type ExpenseFilter } from "../libs/expenseFilter"
 
 const isValidDateKey = (s: string): boolean => /^\d{4}-\d{2}-\d{2}$/.test(s)
