+++ frontend/src/expense/pages/TopPage.tsx
@@ -5,6 +5,7 @@ import { api } from "../../common/libs/api"
 import { categoryGlyph } from "../../common/libs/category"
 import { getErrorMessage } from "../../common/libs/client"
 import type { ExpenseResponse } from "../../common/libs/types"
+import { TopHeatmap } from "../components/TopHeatmap"
 
 const pad = (n: number) => String(n).padStart(2, "0")
 
@@ -21,69 +22,6 @@ const dayShort = (iso: string): string => {
 const timeLabel = (iso: string): string => {
   const d = new Date(iso)
   return `${pad(d.getHours())}:${pad(d.getMinutes())}`
-}
-
-interface HeatmapProps {
-  year: number
-  month: number
-  today: number
-  totals: number[]
-  onSelectDay: (day: number) => void
-}
-
-const Heatmap = ({ year, month, today, totals, onSelectDay }: HeatmapProps) => {
-  const daysInMonth = new Date(year, month, 0).getDate()
-  const firstDow = new Date(year, month - 1, 1).getDay()
-  const max = Math.max(...totals, 0)
-
-  return (
-    <div>
-      <div className="mb-1.5 grid grid-cols-7 text-center text-[9px] font-semibold tracking-wide text-gray-400 dark:text-gray-500">
-        {["S", "M", "T", "W", "T", "F", "S"].map((d, i) => (
-          <span key={i}>{d}</span>
-        ))}
-      </div>
-      <div className="grid grid-cols-7 gap-1">
-        {Array.from({ length: firstDow }).map((_, i) => (
-          <div key={`sp${i}`} />
-        ))}
-        {Array.from({ length: daysInMonth }).map((_, i) => {
-          const day = i + 1
-          const v = totals[i] ?? 0
-          const isToday = day === today
-          const isFuture = day > today
-          const intensity = max > 0 ? 22 + (v / max) * 72 : 0
-          const cellStyle =
-            !isFuture && v > 0
-              ? {
-                  background: `color-mix(in oklab, var(--color-primary) ${intensity}%, #edf2f7)`,
-                  color: v / max > 0.5 ? "#fff" : "#4a5568",
-                }
-              : undefined
-          return (
-            <button
-              key={i}
-              type="button"
-              onClick={() => onSelectDay(day)}
-              disabled={isFuture}
-              style={cellStyle}
-              className={`flex aspect-square items-center justify-center rounded-md text-[10px] font-semibold disabled:cursor-default ${
-                isFuture
-                  ? "border border-dashed border-gray-200 text-gray-300 dark:border-gray-700 dark:text-gray-600"
-                  : isToday
-                    ? "border-[1.5px] border-gray-900 dark:border-gray-100"
-                    : v === 0
-                      ? "bg-gray-100 text-gray-400 dark:bg-gray-700 dark:text-gray-500"
-                      : ""
-              }`}
-            >
-              {day}
-            </button>
-          )
-        })}
-      </div>
-    </div>
-  )
 }
 
 export const TopPage = () => {
@@ -142,7 +80,7 @@ export const TopPage = () => {
       </div>
 
       <div className="shrink-0 pt-4">
-        <Heatmap
+        <TopHeatmap
           year={year}
           month={month}
           today={today}
